requirements.txt
discord.py
praw
PyNaCl
deep_translator
wikipedia
lxml